requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
streamlit==1.28.0
Pillow==10.1.0
numpy==1.24.3
